Move FindAnEvent and SelectAnEvent pages to earlier in grant application flow
